// frontend/src/app/api/model/tokenObtainPair.ts
/**
 * ChemFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TokenObtainPair { 
    username: string;
    password: string;
    readonly access: string;
    readonly refresh: string;
}